requirements.txt
torch==1.2.0
torchvision==0.4.0
pyyaml==5.3.1
pillow==6.2.1
future==0.18.2
tensorboard==2.2.0
pandas==1.0.3